edits and bug fixes
Made a lot of changes here - a few smaller ones just to make the code run in my environment but some larger ones too. For example, the variable name "data" was causing problems so I changed it to clean_data everywhere - some function arguments didn't have speech marks either which was causing issues. Also added return statements to the functions which fixed the issues there. Still a lot of work to do though.